# openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "electrostore",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "description": "URL de base personnalisable",
      "variables": {
        "baseUrl": {
          "default": "https://exemple.com",
          "description": "URL de base personnalisée",
          "examples": [
            "https://api.exemple.com",
            "https://staging-api.exemple.com",
            "http://localhost:8080"
          ]
        }
      }
    }
  ],
  "host": "{baseUrl}",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/item/{id_item}/box/{id_box}": {
      "get": {
        "tags": [
          "ItemBox"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ItemBox"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemBoxByItemDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemBoxByItemDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemBoxByItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ItemBox"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemBoxDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemBoxDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemBoxDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadItemBoxDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ItemBox"
        ],
        "parameters": [
          {
            "name": "id_item",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/store/{id_store}/box": {
      "get": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of items",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadBoxDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadBoxDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadBoxDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoxByStoreDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoxByStoreDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoxByStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/{id_store}/box/{id_box}": {
      "get": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoxByStoreDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoxByStoreDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoxByStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/store/{id_store}/box/{id_box}/show": {
      "post": {
        "tags": [
          "StoreBox"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "red",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "green",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "blue",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "timeshow",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "animation",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/store/{id_store}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/tag/{id_tag}/box/{id_box}": {
      "get": {
        "tags": [
          "TagBox"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TagBox"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadBoxTagDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TagBox"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_box",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/tag/{id_tag}/store/{id_store}": {
      "get": {
        "tags": [
          "TagStore"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TagStore"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStoreTagDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TagStore"
        ],
        "parameters": [
          {
            "name": "id_tag",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id_store",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateBoxByStoreDto": {
        "required": [
          "xend_box",
          "xstart_box",
          "yend_box",
          "ystart_box"
        ],
        "type": "object",
        "properties": {
          "xstart_box": {
            "type": "integer",
            "format": "int32"
          },
          "ystart_box": {
            "type": "integer",
            "format": "int32"
          },
          "xend_box": {
            "type": "integer",
            "format": "int32"
          },
          "yend_box": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreateItemBoxByItemDto": {
        "required": [
          "qte_item_box",
          "seuil_max_item_item_box"
        ],
        "type": "object",
        "properties": {
          "qte_item_box": {
            "type": "integer",
            "format": "int32"
          },
          "seuil_max_item_item_box": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ReadBoxDto": {
        "type": "object",
        "properties": {
          "id_box": {
            "type": "integer",
            "format": "int32"
          },
          "xstart_box": {
            "type": "integer",
            "format": "int32"
          },
          "ystart_box": {
            "type": "integer",
            "format": "int32"
          },
          "xend_box": {
            "type": "integer",
            "format": "int32"
          },
          "yend_box": {
            "type": "integer",
            "format": "int32"
          },
          "id_store": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ReadBoxTagDto": {
        "type": "object",
        "properties": {
          "id_tag": {
            "type": "integer",
            "format": "int32"
          },
          "id_box": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ReadItemBoxDto": {
        "type": "object",
        "properties": {
          "id_box": {
            "type": "integer",
            "format": "int32"
          },
          "id_item": {
            "type": "integer",
            "format": "int32"
          },
          "qte_item_box": {
            "type": "integer",
            "format": "int32"
          },
          "seuil_max_item_item_box": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ReadStoreDto": {
        "type": "object",
        "properties": {
          "id_store": {
            "type": "integer",
            "format": "int32"
          },
          "nom_store": {
            "type": "string",
            "nullable": true
          },
          "xlength_store": {
            "type": "integer",
            "format": "int32"
          },
          "ylength_store": {
            "type": "integer",
            "format": "int32"
          },
          "mqtt_name_store": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReadStoreTagDto": {
        "type": "object",
        "properties": {
          "id_store": {
            "type": "integer",
            "format": "int32"
          },
          "id_tag": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateBoxByStoreDto": {
        "type": "object",
        "properties": {
          "xstart_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ystart_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "xend_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "yend_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateItemBoxDto": {
        "type": "object",
        "properties": {
          "new_id_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "qte_item_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "seuil_max_item_item_box": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateStoreDto": {
        "type": "object",
        "properties": {
          "nom_store": {
            "type": "string",
            "nullable": true
          },
          "xlength_store": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ylength_store": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "mqtt_name_store": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [
        
      ]
    }
  ]
}